pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "aa-wizardindustry-plugin"
description = "Wizard Industry"
readme = "README.md"
keywords = [
    "allianceauth",
    "wizard",
]
license = { file = "LICENSE" }
authors = [
    { name = "Dusty Meg", email = "aa_plugins@mmoore.uk" },
]
requires-python = ">=3.10"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dynamic = [
    "version",
]
dependencies = [
    "allianceauth<5,>=4",
    "django-eveuniverse>=1.5.4",
    "aa-blueprints>=1.7.0",
]
urls.Changelog = "https://github.com/Dusty-Meg/aa-wizardindustry-plugin/blob/master/CHANGELOG.md"
urls."Issue / Bug Reports" = "https://github.com/Dusty-Meg/aa-wizardindustry-plugin/issues"

[tool.hatch.version]
path = "wizardindustry/__init__.py"

[tool.hatch.build]
include = [
    "/wizardindustry",
]
